separate tools from resources
